fix group validation from failing when checking powershell files
This commit addresses a bug found when group task performs task
validation. The final step of validation is to ensure the powershell
files exist. Tasks may have multiple args such as powershell_vars. The
code was getting key error when trying to check powershell_vars when
checking powershell files. An additional check was added to ensure only
powershell files are being checked for this step.