convert class creating paws logger into a mixin class type
This commit transitions the existing logging class into a mixin type
class. This gives us the ability to inherit the class in our main paws
class and paws task classes. Allowing us to access paws logger by a
class attribute (self.logger..).

This can be expanded to other classes in a future commit.