Modifying pylint keys default values
This commit increases the maximum number of lines a function/method
can have. Default is 50 and is increased to 80. It also increases the
number of similar lines from 5 to 100. This was causing pylint to state
the similarity between the provision, show and teardown tasks. An issue
already exists to move these into a common module. Once that is
complete, this default value can be switched back.