Provide more information when fetching paws version (#5)
Originally when running paws --version it would display the current version installed. This patch provides more information to users. It
will provide the installed version, the latest version available and a
message. Now when newer versions of paws are available, it will give
users a notification that they could upgrade their installed version.