lock paramiko to set version to resolve gssapi incompatibility (#11)
This will resolve the bug found when a user has the gssapi library
installed and then tries to run paws. When they go to run paws, it
will fail with a paramiko exception. Rolling back the version resolves
this bug. To note: paws does not require gssapi. We will upgrade to
the latest version of paramiko once they fix the bug.

More information can be referenced in #9