use readthedocs over github pages
Read the docs provides the ability for documentation per versions. This
is nice when a user sticks with a certain version of paws, they can
continue to see the docuemntation for that version. Example would be if
paws changes all parameter inputs, you can see the prior docs has the
original parameter input for paws.