remove python-novaclient dependency
Before paws required nova to get the admin password and use that
password for authentication. It was also using authentication by SSH
keys. Then the user was able to define the administrator account to be
used for configuring windows services. By default now paws will require
the administrator password key for resources to be defined. This account
will be the one used to login to windows systems by ansible. Any SSH
connections by paramiko library will use the private key.

This also reduces the number of package dependencies paws requires.