bug fix to resolve paws hanging when releasing ipv4 connections (#23)
This commit resolves a bug found when trying to create snapshots from
a configured Windows system. Before creating the snapshot, paws releases
all ipv4 connections to ensure networking will work when a new instance
is created based on the snapshot. The command needs to be a fire and
forget call, since it looses network connectivity immediately. A fix
is in place to fire and forget the command.