ability to state attempts and delay when checking snapshot building (#35)
By default when paws goes to take a snapshot of an active instance, it
has a default number of attempts (30) and a default time to wait (20s)
between attempts. There is cases where these default values are
exceeded. This commit allows the user to define custom values for these
when checking if the snapshot build is complete.