remove force upgrading ansible part of post install task
Previously when paws was installed by rpm, a post task was called
to use pip to install paws required python packages. This was always
forcing ansible to be upgraded on the system (even if the system had
a version already installed). This commit removes forcing the upgrade
of ansible and will use the version of ansible installed on the
system where paw is installed. Or it will install the latest verison
from the repositories available.

To note: this commit does not affect pip installs. Pip installs will
continue to get the latest version of ansible installed.