ability to override the default attempts for instance building
Previously paws would only do 30 attempts checking if an instance is
successfully provisioned. There is use cases where this number of
attempts is not enough and needs to be more. With this change, users
can now supply their own number of attempts when checking if an instance
is provisioned. Each resource can have their own unique number of
attempts.